backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Browser-Use Chat Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]",
    "uvicorn[standard]",
    "sqlalchemy>=2.0",
    "aiosqlite",
    "greenlet",
    "pydantic-settings",
    "python-dotenv",
    "openai",
    "browser-use",
]

[tool.setuptools.packages.find]
include = ["app*", "worker*"]
